requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
Django==4.2.6
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
Markdown==3.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
